requirements.txt
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.38
python-dotenv>=1.0.0
requests>=2.31.0
APScheduler>=3.10.4
pytz>=2024.1
